Create updater.vbs
This is the malicious vba that builds the other scripts. I've commented out the script building portions, but kept the schtask portions. I've also gotten rid of the dumb macro calls and some other things so you can actually run it, but you really shouldn't.  If you run it, you can get rid of what happens by removing the schtask.