Update Script1.ps1
Got the names mixed up.  This script grabs commands from the C2 and then sends them to temp.ps1 to be ran.  This version won't run the malware. But is written for so it's easily read.  Much of the contents can be changed to literally anything else and work. So the whole yara thing may not work for more than a day for this specific malware.